ref!: Skip by default overwriting duplicates
The previous behavior for the import cmd was to overwrite any existing
repository. As it was explained in #9, this can be quite destructive as
it deletes the duplicated directory removing any local changes the user
my want to keep.

To provide some form of safe-rails for those users that do not expect
such behavior, I replaced the `--skip` flag for a `--force` and to
indicate that those repositories that are duplicated should be
overwritten.