ref: Improve handling of repository Path
This functionality was originally introduced for the switch command. It
basically enables the user to provide the directory name and as long as
a directory with that name is found from the current path, it will be
used as root.

This is a neat addition to save time writing the complete path to the
repository to be handled.

Additionally, it does not break the previous behavior where if a full
path is provided, that one will be used. Similarly, if no path is
provided, the current path will be taken as root for most commands.